Create minors components to simplify the code
